Migración de IDs de int a long en toda la solución
Se reemplazan todos los identificadores de tipo int por long en entidades, DTOs, comandos, queries, viewmodels y servicios de cifrado. Se actualizan controladores, lógica de acceso a datos, colecciones y tests para soportar el nuevo tipo. Este cambio amplía la capacidad de la aplicación para manejar un mayor rango de entidades.

+++ src/SFManagement.Application/Abstractions/IIdEncryptionService.cs
@@ -2,7 +2,7 @@ namespace SFManagement.Application.Abstractions;
 
 public interface IIdEncryptionService
 {
-    string Encrypt(int id);
-    int Decrypt(string encrypted);
-    bool TryDecrypt(string encrypted, out int id);
+    string Encrypt(long id);
+    long Decrypt(string encrypted);
+    bool TryDecrypt(string encrypted, out long id);
 }
