+++ src/SFManagement.Application/Commands/EvaluateTaskCommand.cs
@@ -1,3 +1,3 @@
-public record EvaluateTaskCommand(int TaskId, int WorkerId, IReadOnlyList<SkillEvaluation> Evaluations);
+public record EvaluateTaskCommand(long TaskId, long WorkerId, IReadOnlyList<SkillEvaluation> Evaluations);
 
 public record SkillEvaluation(int SkillPosition, double BasePoints);
