+++ src/SFManagement.Application/Commands/UpdateTaskCommand.cs
@@ -3,8 +3,8 @@ using SFManagement.Domain.Enums;
 namespace SFManagement.Application.Commands;
 
 public record UpdateTaskCommand(
-    int TaskId,
-    int ProjectId,
+    long TaskId,
+    long ProjectId,
     string Title,
     string? Description,
     Criticality Criticality,
