+++ src/SFManagement.Application/Commands/UpdateWorkerCommand.cs
@@ -1,4 +1,4 @@
 namespace SFManagement.Application.Commands;
 
-public record UpdateWorkerCommand(int WorkerId, string Name, string Role, float[]? SkillsVector = null);
+public record UpdateWorkerCommand(long WorkerId, string Name, string Role, float[]? SkillsVector = null);
