+++ src/SFManagement.Application/Queries/GetWorkersNotInProjectQuery.cs
@@ -2,7 +2,7 @@ using SFManagement.Application.DTOs;
 
 namespace SFManagement.Application.Queries;
 
-public record GetWorkersNotInProjectQuery(int ProjectId);
+public record GetWorkersNotInProjectQuery(long ProjectId);
 
 public interface IGetWorkersNotInProjectQueryHandler
 {
